+++ vpin-connector-vps/src/main/java/de/mephisto/vpin/connectors/vps/VpsChangeListener.java
@@ -1,0 +1,9 @@
+package de.mephisto.vpin.connectors.vps;
+
+import de.mephisto.vpin.connectors.vps.model.VpsTableDiff;
+
+import java.util.List;
+
+public interface VpsChangeListener {
+  void vpsSheetChanged(List<VpsTableDiff> diff);
+}