+++ RELEASE_NOTES.md
@@ -3,4 +3,5 @@
 ### Bugfixes
 
 - Fixed offset of table rows: I had to change the selected row styling for this again.
-- Improved exception handling for failed updates.
+- Improved exception handling for failed updates
+- Fixed issue that the VPS sheet has been downloaded, but the data has not been refreshed (until next start).